requirements.txt
PyAutoGUI>=0.9.52
pyxdg>=0.27
Pillow>=8.1.0